+++ cypress/elements/optionsModal/axes.js
@@ -1,17 +1,21 @@
-import { clickCheckbox, typeInput } from '../common'
+import { typeInput } from '../common'
 
-const titleCheckboxEl = 'axis-title-checkbox'
 const titleInputEl = 'axis-title-input'
 const rangeMinInputEl = 'axis-range-min-input'
 const rangeMaxInputEl = 'axis-range-max-input'
+const axisTitleRadiosEl = 'option-chart-title'
+const verticalTabsEl = 'axes-tabs'
 
 const getAxisSelector = (axis, selector) => `${axis}-${selector}`
 
-export const enableAxisTitle = axis =>
-    clickCheckbox(getAxisSelector(axis, titleCheckboxEl))
+export const switchAxesTabTo = tab =>
+    cy.getBySel(verticalTabsEl).containsExact(tab).click()
 
-export const setAxisTitle = (axis, text) =>
+export const setAxisTitleText = (axis, text) =>
     typeInput(getAxisSelector(axis, titleInputEl), text)
+
+export const setAxisTitleToCustom = () =>
+    cy.getBySel(axisTitleRadiosEl).contains('Custom').click()
 
 export const expectAxisTitleToBeValue = (axis, value) =>
     cy