+++ cypress/integration/dimensions/data.cy.js
@@ -1,4 +1,5 @@
 import { DIMENSION_ID_DATA } from '@dhis2/analytics'
+
 import {
     clickDimensionModalHideButton,
     expectDimensionModalToContain,
@@ -6,20 +7,18 @@ import {
     unselectItemByDoubleClick,
     selectItemByDoubleClick,
     expectDataDimensionModalToBeVisible,
-    expectDataItemToBeSelected,
+    expectItemToBeSelected,
     expectDataTypeToBe,
     expectGroupSelectToNotBeVisible,
     expectNoDataItemsToBeSelected,
-    selectAllDataItems,
-    unselectAllDataItems,
-    expectDataItemsSelectedAmountToBeLeast,
-    expectDataItemsSelectedAmountToBe,
-    expectDataItemToBeSelectable,
-    expectDataItemsSelectableAmountToBe,
+    expectSelectedItemsAmountToBeLeast,
+    expectSelectedItemsAmountToBe,
+    expectItemToBeSelectable,
+    expectSelectableItemsAmountToBe,
     inputSearchTerm,
     switchDataTypeTo,
     clearSearchTerm,
-    expectDataItemsSelectableAmountToBeLeast,
+    expectSelectableItemsAmountToBeLeast,
     expectGroupSelectToBeVisible,
     switchGroupTo,
     selectFirstDataItem,
@@ -33,16 +32,20 @@ import {
     expectSubGroupSelectToBeVisible,
     expectSubGroupSelectToBe,
     switchSubGroupTo,
-} from '../elements/dimensionModal'
-import { openDimension } from '../elements/dimensionsPanel'
-import { goToStartPage } from '../elements/startScreen'
+    expectSourceToNotBeLoading,
+    unselectAllItemsByButton,
+    selectAllItemsByButton,
+} from '../../elements/dimensionModal'
+import { openDimension } from '../../elements/dimensionsPanel'
+import { goToStartPage } from '../../elements/startScreen'
 import {
     TEST_DATA_ELEMENTS,
     TEST_DATA_SETS,
     TEST_INDICATORS,
-} from '../utils/data'
+} from '../../utils/data'
 
 const PAGE_SIZE = 50
+const DATA_ITEMS_URL = '/dataItems'
 
 describe('Data dimension', () => {
     describe('initial state', () => {
@@ -50,9 +53,9 @@ describe('Data dimension', () => {
             goToStartPage()
         })
         it('opens the data dimension modal', () => {
-            cy.intercept('GET', '/dataItems').as('dataItems')
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
             openDimension(DIMENSION_ID_DATA)
-            cy.wait('@dataItems').then(({ request, response }) => {
+            cy.wait('@request').then(({ request, response }) => {
                 expect(request.url).to.contain('page=1')
                 expect(response.statusCode).to.eq(200)
                 expect(response.body.dataItems.length).to.eq(PAGE_SIZE)
@@ -74,7 +77,7 @@ describe('Data dimension', () => {
         const firstPageItemName = TEST_INDICATORS[0].name
         it('an item can be selected by double click', () => {
             selectItemByDoubleClick(firstPageItemName)
-            expectDataItemToBeSelected(firstPageItemName)
+            expectItemToBeSelected(firstPageItemName)
         })
         it('an item can be unselected by double click', () => {
             unselectItemByDoubleClick(firstPageItemName)
@@ -82,7 +85,7 @@ describe('Data dimension', () => {
         })
         it('an item can be selected by button', () => {
             selectItemByButton(firstPageItemName)
-            expectDataItemToBeSelected(firstPageItemName)
+            expectItemToBeSelected(firstPageItemName)
         })
         it('an item can be unselected by button', () => {
             unselectItemByButton(firstPageItemName)
@@ -92,60 +95,107 @@ describe('Data dimension', () => {
     describe('selecting all and fetching more', () => {
         const secondPageItemName = 'BCG doses'
         it('all items can be selected', () => {
-            cy.intercept('GET', '/dataItems').as('dataItems')
-            selectAllDataItems()
-            expectDataItemsSelectedAmountToBeLeast(PAGE_SIZE)
-            cy.wait('@dataItems').then(({ request, response }) => {
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
+            selectAllItemsByButton()
+            expectSelectedItemsAmountToBeLeast(PAGE_SIZE)
+            cy.wait('@request').then(({ request, response }) => {
                 expect(request.url).to.contain('page=2')
                 expect(response.statusCode).to.eq(200)
                 expect(response.body.dataItems.length).to.eq(PAGE_SIZE)
             })
+            expectSourceToNotBeLoading()
         })
         it('more items are fetched', () => {
-            expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
-            expectDataItemToBeSelectable(secondPageItemName)
+            expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
+            expectItemToBeSelectable(secondPageItemName)
         })
         it('all items can be unselected', () => {
-            unselectAllDataItems()
+            unselectAllItemsByButton()
             expectNoDataItemsToBeSelected()
         })
         it('more items are fetched when scrolling down', () => {
-            cy.intercept('GET', '/dataItems').as('request')
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
             scrollSourceToBottom()
             cy.wait('@request').then(({ request, response }) => {
                 expect(request.url).to.contain('page=3')
                 expect(response.statusCode).to.eq(200)
                 expect(response.body.dataItems.length).to.eq(PAGE_SIZE)
             })
-            expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE * 3)
+            expectSourceToNotBeLoading()
+            expectSelectableItemsAmountToBeLeast(PAGE_SIZE * 3)
         })
     })
     describe('global search', () => {
         const testSearchTerm = 'Dispenser' // Use a data element for the third step to work
-        it('recieves a search term', () => inputSearchTerm(testSearchTerm))
+        it('recieves a search term', () => {
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
+            inputSearchTerm(testSearchTerm)
+            cy.wait('@request').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(request.url).to.contain(testSearchTerm)
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataItems.length).to.eq(1)
+            })
+            expectSourceToNotBeLoading()
+        })
         // TODO: Test that the search is only called once, i.e. debounce works
         it('search result is displayed', () => {
-            expectDataItemsSelectableAmountToBe(1)
-            expectDataItemToBeSelectable(testSearchTerm)
+            expectSelectableItemsAmountToBe(1)
+            expectItemToBeSelectable(testSearchTerm)
         })
         it('search result is maintained when switching data type', () => {
+            cy.intercept('GET', '/dataElements').as('dataElements')
             switchDataTypeTo('Data elements')
-            expectDataItemsSelectableAmountToBe(1)
-            expectDataItemToBeSelectable(testSearchTerm)
+            cy.wait('@dataElements').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataElements.length).to.be.eq(1)
+            })
+            expectSourceToNotBeLoading()
+            expectSelectableItemsAmountToBe(1)
+            expectItemToBeSelectable(testSearchTerm)
             clearSearchTerm()
+            cy.wait('@dataElements').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataElements.length).to.be.eq(PAGE_SIZE)
+            })
+            expectSourceToNotBeLoading()
+            cy.intercept('GET', DATA_ITEMS_URL).as('dataItems')
             switchDataTypeToAll()
-            expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
+            cy.wait('@dataItems').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataItems.length).to.eq(PAGE_SIZE)
+            })
+            expectSourceToNotBeLoading()
+            expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
         })
         it('search displays a correct error message', () => {
             const testSearchTermWithNoMatch = 'nomatch'
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
             inputSearchTerm(testSearchTermWithNoMatch)
+            cy.wait('@request').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(request.url).to.contain(testSearchTermWithNoMatch)
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataItems.length).to.eq(0)
+            })
+            expectSourceToNotBeLoading()
             expectEmptySourceMessageToBe(
                 `Nothing found for "${testSearchTermWithNoMatch}"`
             )
         })
         it('search result can be cleared', () => {
+            cy.intercept('GET', DATA_ITEMS_URL).as('request')
             clearSearchTerm()
-            expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
+            cy.wait('@request').then(({ request, response }) => {
+                expect(request.url).to.contain('page=1')
+                expect(response.statusCode).to.eq(200)
+                expect(response.body.dataItems.length).to.be.eq(PAGE_SIZE)
+            })
+            expectSourceToNotBeLoading()
+            expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
         })
         it('modal is closed', () => {
             clickDimensionModalHideButton()
@@ -167,7 +217,15 @@ describe('Data dimension', () => {
         {
             name: 'Data elements',
             testGroup: { name: 'Measles', itemAmount: 3 },
-            testSubGroup: { name: 'Details only', itemAmount: '10' },
+            testSubGroup: {
+                name: 'Details only',
+                itemAmount: 10,
+                endpoint: {
+                    hasMultiplePages: true,
+                    requestUrl: '/dataElementOperands',
+                    responseBody: 'dataElementOperands',
+                },
+            },
             testItem: { name: TEST_DATA_ELEMENTS[2].name },
             defaultGroup: { name: 'All groups' },
             defaultSubGroup: { name: 'Totals only' },
@@ -197,7 +255,7 @@ describe('Data dimension', () => {
             defaultGroup: { name: 'All programs' },
             endpoint: {
                 hasMultiplePages: true,
-                requestUrl: '/dataItems',
+                requestUrl: DATA_ITEMS_URL,
                 responseBody: 'dataItems',
             },
         },
@@ -208,7 +266,7 @@ describe('Data dimension', () => {
             defaultGroup: { name: 'All programs' },
             endpoint: {
                 hasMultiplePages: true,
-                requestUrl: '/dataItems',
+                requestUrl: DATA_ITEMS_URL,
                 responseBody: 'dataItems',
             },
         },
@@ -220,8 +278,18 @@ describe('Data dimension', () => {
                 expectDataDimensionModalToBeVisible()
             })
             it(`switches to ${testDataType.name}`, () => {
+                cy.intercept('GET', testDataType.endpoint.requestUrl).as(
+                    'request'
+                )
                 switchDataTypeTo(testDataType.name)
-                expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
+                cy.wait('@request').then(({ response }) => {
+                    expect(response.statusCode).to.eq(200)
+                    expect(
+                        response.body[testDataType.endpoint.responseBody].length
+                    ).to.be.least(1)
+                })
+                expectSourceToNotBeLoading()
+                expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
             })
             it('group select is visible', () => {
                 expectGroupSelectToBeVisible()
@@ -241,25 +309,34 @@ describe('Data dimension', () => {
                                 .length
                         ).to.be.least(1)
                     })
-                    expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE + 1)
+                    expectSourceToNotBeLoading()
+                    expectSelectableItemsAmountToBeLeast(PAGE_SIZE + 1)
                 })
             }
             it('an item can be selected', () => {
                 selectItemByDoubleClick(testDataType.testItem.name)
-                expectDataItemToBeSelected(testDataType.testItem.name)
+                expectItemToBeSelected(testDataType.testItem.name)
             })
             it(`group can be changed to "${testDataType.testGroup.name}"`, () => {
+                cy.intercept('GET', testDataType.endpoint.requestUrl).as(
+                    'request'
+                )
                 switchGroupTo(testDataType.testGroup.name)
+                cy.wait('@request').then(({ request, response }) => {
+                    expect(request.url).to.contain('page=1')
+                    expect(response.statusCode).to.eq(200)
+                })
+                expectSourceToNotBeLoading()
                 expectGroupSelectToBe(testDataType.testGroup.name)
-                expectDataItemsSelectableAmountToBe(
+                expectSelectableItemsAmountToBe(
                     testDataType.testGroup.itemAmount
                 )
-                expectDataItemToBeSelected(testDataType.testItem.name)
+                expectItemToBeSelected(testDataType.testItem.name)
             })
             it('the first item can be selected', () => {
                 selectFirstDataItem()
-                expectDataItemsSelectedAmountToBe(2)
-                expectDataItemsSelectableAmountToBe(
+                expectSelectedItemsAmountToBe(2)
+                expectSelectableItemsAmountToBe(
                     testDataType.testGroup.itemAmount - 1
                 )
             })
@@ -270,40 +347,103 @@ describe('Data dimension', () => {
                 })
 
                 it(`sub group can be changed to "${testDataType.testSubGroup.name}"`, () => {
+                    cy.intercept(
+                        'GET',
+                        testDataType.testSubGroup.endpoint?.requestUrl ||
+                            testDataType.endpoint.requestUrl
+                    ).as('request')
                     switchSubGroupTo(testDataType.testSubGroup.name)
+                    cy.wait('@request').then(({ request, response }) => {
+                        expect(request.url).to.contain('page=1')
+                        expect(response.statusCode).to.eq(200)
+                        expect(
+                            response.body[
+                                testDataType.testSubGroup.endpoint
+                                    ?.responseBody ||
+                                    testDataType.endpoint.responseBody
+                            ].length
+                        ).to.be.eq(testDataType.testSubGroup.itemAmount)
+                    })
+                    expectSourceToNotBeLoading()
                     expectSubGroupSelectToBe(testDataType.testSubGroup.name)
-                    expectDataItemsSelectableAmountToBe(
+                    expectSelectableItemsAmountToBe(
                         testDataType.testSubGroup.itemAmount
                     )
-                    expectDataItemToBeSelected(testDataType.testItem.name)
+                    expectItemToBeSelected(testDataType.testItem.name)
                 })
                 it(`sub group can be changed back to "${testDataType.defaultSubGroup.name}"`, () => {
+                    cy.intercept('GET', testDataType.endpoint.requestUrl).as(
+                        'request'
+                    )
                     switchSubGroupTo(testDataType.defaultSubGroup.name)
+                    cy.wait('@request').then(({ request, response }) => {
+                        expect(request.url).to.contain('page=1')
+                        expect(response.statusCode).to.eq(200)
+                    })
+                    expectSourceToNotBeLoading()
                     expectSubGroupSelectToBe(testDataType.defaultSubGroup.name)
-                    expectDataItemsSelectableAmountToBe(
+                    expectSelectableItemsAmountToBe(
                         testDataType.testGroup.itemAmount - 1
                     )
-                    expectDataItemToBeSelected(testDataType.testItem.name)
+                    expectItemToBeSelected(testDataType.testItem.name)
                 })
             }
             it('search displays a correct error message', () => {
                 const testSearchTermWithNoMatch = 'nomatch'
+                cy.intercept('GET', testDataType.endpoint.requestUrl).as(
+                    'request'
+                )
                 inputSearchTerm(testSearchTermWithNoMatch)
+                cy.wait('@request').then(({ request, response }) => {
+                    expect(request.url).to.contain('page=1')
+                    expect(request.url).to.contain(testSearchTermWithNoMatch)
+                    expect(response.statusCode).to.eq(200)
+                    expect(
+                        response.body[testDataType.endpoint.responseBody].length
+                    ).to.eq(0)
+                })
+                expectSourceToNotBeLoading()
                 expectEmptySourceMessageToBe(
                     `No ${testDataType.name.toLowerCase()} found for "${testSearchTermWithNoMatch}"`
                 )
             })
             it('selection and filter can be reset', () => {
-                unselectAllDataItems()
+                unselectAllItemsByButton()
                 expectNoDataItemsToBeSelected()
+                cy.intercept('GET', testDataType.endpoint.requestUrl).as(
+                    testDataType.endpoint.requestUrl
+                )
                 clearSearchTerm()
-                expectDataItemsSelectableAmountToBe(
+                cy.wait(`@${testDataType.endpoint.requestUrl}`).then(
+                    ({ request, response }) => {
+                        expect(request.url).to.contain('page=1')
+                        expect(response.statusCode).to.eq(200)
+                    }
+                )
+                expectSourceToNotBeLoading()
+                expectSelectableItemsAmountToBe(
                     testDataType.testGroup.itemAmount
                 )
                 switchGroupToAll()
-                expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
+                cy.wait(`@${testDataType.endpoint.requestUrl}`).then(
+                    ({ request, response }) => {
+                        expect(request.url).to.contain('page=1')
+                        expect(response.statusCode).to.eq(200)
+                    }
+                )
+                expectSourceToNotBeLoading()
+                expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
+                if (testDataType.endpoint.requestUrl !== DATA_ITEMS_URL) {
+                    cy.intercept('GET', DATA_ITEMS_URL).as('/dataItems')
+                }
                 switchDataTypeToAll()
-                expectDataItemsSelectableAmountToBeLeast(PAGE_SIZE)
+                cy.wait('@/dataItems').then(({ request, response }) => {
+                    expect(request.url).to.contain('page=1')
+                    expect(response.statusCode).to.eq(200)
+                    expect(response.body.dataItems.length).to.eq(PAGE_SIZE)
+                })
+                expectSourceToNotBeLoading()
+                expectSelectableItemsAmountToBeLeast(PAGE_SIZE)
             })
             it('modal is closed', () => {
                 clickDimensionModalHideButton()